Popularity

+++ objects/likes.js
@@ -1,23 +1,40 @@
 let update = require('./update.js');
 
-class Likes{
-
-    static handleLikes(data, socket, db, sess, allUsers){
+class Likes {
+    static async handleLikes(data, socket, db, sess, allUsers, io){
+        const block = sess.data.block && data.login && sess.data.block.indexOf(data.login.id) !== -1;
+        const sql = "SELECT JSON_CONTAINS((SELECT list FROM block WHERE userid=?), ?, '$') AS blocked";
+        let blocked = false;
+        if (data.login) {
+            [blocked] = await db.execute(sql, [data.login.id, `${sess.data.id}`]);
+        }
 
         switch(data.type){
             case 'Add':
-                Likes.addLike(data.login, socket, db, sess);
-                Likes.sendNotif(db, data.login, socket, allUsers);
+                if (!block) {
+                    await Likes.addLike(data.login, socket, db, sess, io, blocked);
+                    if (!blocked || blocked[0].blocked !== 1) {
+                        Likes.sendNotif(db, data.login, socket, allUsers);
+                    }
+                }
                 break;
             case 'Remove':
-                Likes.removeLike(data.login, socket, db, sess);
-                Likes.sendNotif(db, data.login, socket, allUsers);
+                if (!block) {
+                    await Likes.removeLike(data.login, socket, db, sess, io, blocked);
+                    if (!blocked || blocked[0].blocked !== 1) {
+                        Likes.sendNotif(db, data.login, socket, allUsers);
+                    }
+                }
                 break ;
             case 'addMatch':
-                Likes.addMatchList(socket, db, sess, data.login, false);
+                if (!block) {
+                    Likes.addMatchList(socket, db, sess, data.login, false);
+                }
                 break;
             case 'delMatch':
-                Likes.deleteMatchList(socket, db, sess, data.login, false);
+                if (!block) {
+                    Likes.deleteMatchList(socket, db, sess, data.login, false);
+                }
                 break ;
             case 'refresh':
                 update.getNotif(db, sess).then(() => socket.emit('user', sess.data));
@@ -26,53 +43,83 @@ class Likes{
 
     }
 
-    static addLike(user, socket, db, sess){
-        let id = sess.data.id;
-        let sql;
+    static async sendProfil(db, sess, socket, data, io, setState){
+        const sql = "SELECT users.id, users.login, users.last, users.first, users.age, users.sexe, users.bio, users.orientation, users.tags, users.spop, users.date, location.city, location.country, location.zipcode, likes.user1, likes.user2, img.imgid, " +
+            "(SELECT st_distance_sphere((SELECT POINT(lon, lat) FROM location WHERE logid = ?), (SELECT POINT(lon, lat) FROM location WHERE logid = ?))) AS distance " +
+            "FROM users LEFT JOIN img ON img.userid = users.id AND img.profil = '1' LEFT JOIN likes ON ((likes.user1 = ? AND likes.user2 = users.id) OR (likes.user1 = users.id AND likes.user2 = ?)) INNER JOIN location ON location.logid = users.id  WHERE users.id = ?";
+        const [rows] = await db.execute(sql, [data.id, sess.data.id, sess.data.id, sess.data.id, data.id]);
 
-        Likes.likeExist(user.id, sess, db)
-            .then((res) => {
-                if (!res)
-                {
-                    sql = "INSERT INTO likes(user1, user2) VALUES (?, ?)";
-                    db.execute(sql, [id, user.id]);
-                    Likes.addNotif(db, sess, 'like', user.id, id);
-                    socket.emit('user', sess.data);
-                } else if (!res.matcha && Number(res.user1) !== id){
-                    Likes.addMatchList(socket, db, sess, user, true);
-                    Likes.addNotif(db, sess, 'match', user.id, id);
-                    sql = "INSERT INTO chat(user1, user2, history) SELECT ?, ? , '[]' " +
-                        "WHERE NOT EXISTS (SELECT user1 FROM chat WHERE (user1= ? AND user2=?) OR (user1=? AND user2=?))"
-                        + " LIMIT 1";
-                    db.execute(sql, [user.id, id, user.id, id, id, user.id])
-                        .catch(e => console.log(e));
-                    sql = "UPDATE likes SET matcha=true WHERE (user1=? AND user2=?) OR (user1=? AND user2=?); " ;
-                    db.execute(sql, [user.id, id, user.id, id]);
-                }
-            })
-            .catch((e) => console.log(e));
+        if (rows[0]){
+            if (setState) {
+                setState(rows[0]);
+            } else {
+                io.emit(data.id, rows[0]);
+            }
+        }
     }
 
-    static removeLike(user, socket, db, sess){
+    static refreshProfil(db, sess, socket, id, io){
+        io.emit(sess.data.id, null);
+        Likes.sendProfil(db, sess, socket, {id: id}, io);
+    }
+
+
+    static addLike(user, socket, db, sess, io, blocked){
         let id = sess.data.id;
         let sql;
 
-        Likes.likeExist(user.id, sess, db)
-            .then((res) =>{
-            if (res && res.matcha){
-                Likes.deleteMatchList(socket, db, sess, user, true);
-                Likes.addNotif(db, sess, 'unmatch', user.id, id);
-                sql = "UPDATE likes SET matcha=false ,user1= (CASE WHEN user1=? THEN ? ELSE ? END)," +
-                    " user2= (CASE WHEN user2=? THEN ? ELSE ? END) WHERE (user1=? AND user2=?) OR (user1=? AND user2=?)";
-                db.execute(sql, [id, user.id, user.id, user.id, id, id, user.id, id, id, user.id]);
-           } else {
-                Likes.addNotif(db, sess, 'unlike', user.id, id);
-                socket.emit('user', sess.data);
-                sql = "DELETE FROM likes WHERE user1=? AND user2=?";
-               db.execute(sql, [id, user.id]);
-           }
-        })
-            .catch((e) => console.log(e));
+            Likes.likeExist(user.id, sess, db)
+                .then(async (res) => {
+                    if (!res) {
+                        sql = "INSERT INTO likes(user1, user2) VALUES (?, ?)";
+                        await db.execute(sql, [id, user.id]);
+                        sql = "UPDATE users SET spop=spop+50 WHERE id = ?";
+                        await db.execute(sql, [user.id]);
+                        Likes.addNotif(db, sess, 'like', user.id, id, blocked);
+                        socket.emit('user', sess.data);
+                    } else if (!res.matcha && Number(res.user1) !== id) {
+                        Likes.addMatchList(socket, db, sess, user, true);
+                        Likes.addNotif(db, sess, 'match', user.id, id, blocked);
+                        sql = "INSERT INTO chat(user1, user2, history) SELECT ?, ? , '[]' " +
+                            "WHERE NOT EXISTS (SELECT user1 FROM chat WHERE (user1= ? AND user2=?) OR (user1=? AND user2=?))"
+                            + " LIMIT 1";
+                        db.execute(sql, [user.id, id, user.id, id, id, user.id])
+                            .catch(e => console.log(e));
+                        sql = "UPDATE likes SET matcha=true WHERE (user1=? AND user2=?) OR (user1=? AND user2=?); ";
+                        await db.execute(sql, [user.id, id, user.id, id]);
+                        sql = "UPDATE users SET spop= (CASE WHEN id = ? THEN spop+50 ELSE spop+100 END) WHERE id = ? OR id = ?";
+                        await db.execute(sql, [id, user.id, id]);
+                    }
+                    Likes.refreshProfil(db, sess, socket, user.id, io);
+                })
+                .catch((e) => console.log(e));
+    }
+
+    static removeLike(user, socket, db, sess, io, blocked){
+        let id = sess.data.id;
+        let sql;
+
+            Likes.likeExist(user.id, sess, db)
+                .then(async (res) => {
+                    if (res && res.matcha) {
+                        Likes.deleteMatchList(socket, db, sess, user, true);
+                        Likes.addNotif(db, sess, 'unmatch', user.id, id, blocked);
+                        sql = "UPDATE users SET spop= (CASE WHEN id = ? THEN spop-50 ELSE spop-100 END) WHERE id = ? OR id = ?";
+                        await db.execute(sql, [id, user.id, id]);
+                        sql = "UPDATE likes SET matcha=false ,user1= (CASE WHEN user1=? THEN ? ELSE ? END)," +
+                            " user2= (CASE WHEN user2=? THEN ? ELSE ? END) WHERE (user1=? AND user2=?) OR (user1=? AND user2=?)";
+                        await db.execute(sql, [id, user.id, user.id, user.id, id, id, user.id, id, id, user.id]);
+                    } else {
+                        Likes.addNotif(db, sess, 'unlike', user.id, id, blocked);
+                        socket.emit('user', sess.data);
+                        sql = "UPDATE users SET spop=(spop-50) WHERE id = ?";
+                        await db.execute(sql, [user.id]);
+                        sql = "DELETE FROM likes WHERE user1=? AND user2=?";
+                        await db.execute(sql, [id, user.id]);
+                    }
+                    Likes.refreshProfil(db, sess, socket, user.id, io);
+                })
+                .catch((e) => console.log(e));
     }
 
 
@@ -118,10 +165,12 @@ class Likes{
         }) ;
     }
 
-    static addNotif(db, sess, type, userId, id){
-        let sql =  "INSERT INTO notif SET type = ?, login = ?, `from` = ?";
+    static addNotif(db, sess, type, userId, id, blocked){
+        if (!blocked || blocked[0].blocked !== 1) {
+            let sql = "INSERT INTO notif SET type = ?, login = ?, `from` = ?";
 
-        db.execute(sql, [type, userId, id]);
+            db.execute(sql, [type, userId, id]);
+        }
     }
 
     static addMatchList(socket, db, sess, user, refresh){
